move types to their packages instead of in streamAccordion
